replace automake with cmake and gcc 7.2
Currently rootserver,updateserver,proxyserver,mergeserver and other
static libraries can be built with cmake, a more powerful and cleaner
build system, and with gcc 7.2. Although there're many not-modern
code resulting in plenty of warnings with -Wall, you can suppress without it.
Those warnings will be solved as soon as possible.

svn_version is added for the moment, later I'll use git to replace it.

Signed-off-by: cshi <baiwfg2@gmail.com>